Added comments for improved code readability

+++ src/cpp/Coordinates/CoordinatesReader.hpp
@@ -2,10 +2,11 @@
 #include <vector>
 #include <tuple>
 
+// The CoordinatesReader class provides method to read coordinates from json file
 class CoordinatesReader {
 public:
     CoordinatesReader(const std::string& filename);
-    std::vector<std::tuple<std::string, int, int, float, int, int, int>> getCoordinates() const;
+    std::vector<std::tuple<std::string, int, int, float, float, int, int>> getCoordinates() const;
 
 private:
     std::string filename;
